pyproject.toml
[tool.poetry]
name = "simple-ticket-system"
version = "0.1.0"
description = ""
authors = ["Michael Ulianich <ulianich_mihail@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
sqlalchemy = "^2.0.30"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
flask-security-too = {extras = ["fsqla"], version = "^5.4.3"}


[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
